\name{caribou16day5indiv}
\alias{caribou16day5indiv}
\docType{data}
\title{
Caribou data set - 5 individuals observed during 16 days
}
\description{
Caribou data set contains 105 observations of  5 individuals observed during 16 days.
}
\usage{data(caribou16day5indiv)}
\format{
  
 
}
\details{

}
\source{

}
\references{

}
\examples{
data(caribou16day5indiv)
}
\keyword{datasets}
