\name{guanaco.reloc.16day.5indiv}
\alias{guanaco.reloc.16day.5indiv}
\docType{data}
\title{
Guanaco data set - 5 individuals observed during 16 days
}
\description{
Caribou data set contains 40 observations of  5 individuals observed during 16 days.
}
\usage{data(guanaco.reloc.16day.5indiv)}
\format{
  The format is:
Formal class 'Individuals' [package "animalmove"] with 6 slots
  ..@ group.by   : chr "pop.type"
  ..@ data       :'data.frame':	40 obs. of  2 variables:
  .. ..$ id      : Factor w/ 5 levels "caro514","Iris474",..: 1 1 1 1 1 1 1 1 2 2 ...
  .. ..$ pop.type: chr [1:40] "guanaco" "guanaco" "guanaco" "guanaco" ...
  ..@ coords.nrs : num(0) 
  ..@ coords     : num [1:40, 1:2] -820509 -805425 -809713 -808318 -774552 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "x" "y"
  ..@ bbox       : num [1:2, 1:2] -820509 -511436 -749951 -454645
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:2] "x" "y"
  .. .. ..$ : chr [1:2] "min" "max"
  ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slots
  .. .. ..@ projargs: chr NA
}
\details{

}
\source{

}
\references{

}
\examples{
data(guanaco.reloc.16day.5indiv)
}
\keyword{datasets}
