\docType{class}
\name{MCIndex}
\alias{MCIndex}
\alias{MCIndex-class}
\title{The MCIndex class}
\description{
The \code{MCIndex} object contains Movement Coordination
Index data frame, which is computed for the populations of
Inidividuals.
}
\note{
Object is constructed as a result of mci.index method call
on the data frame of the Individuals
}
\section{Slots}{
  \describe{ \item{\code{data}:}{Object of class
  \code{"data.frame"}, columns: id - time lag, pop.type -
  population type, mci.index - MCI value} }
}
\author{
Irina Belyaeva
}

