\docType{methods}
\name{pdi.index,Individuals-method}
\alias{pdi.index,Individuals-method}
\title{Computes Population Dispersion Index for the population}
\usage{
\S4method{pdi.index}{Individuals}(object, scale, percent = 95, unin = c("m",
  "km"), unout = c("ha", "km2", "m2"), ...)
}
\arguments{
  \item{object}{An Individual object that contains species
  from the same population}
}
\description{
Computes Population Dispersion Index for the population
}

