\docType{methods}
\name{polygon,PDIndex,missing-method}
\alias{polygon,PDIndex,missing-method}
\title{Plots Population Dispersion Index Object}
\usage{
\S4method{polygon}{PDIndex,missing}(x, y, ...)
}
\arguments{
  \item{x}{vectors containing the coordinates of the
  vertices of the polygon.}

  \item{...}{graphical parameters such as \code{xpd},
  \code{lend}, \code{ljoin} and \code{lmitre} can be given
  as arguments.}
}
\description{
Plots Population Dispersion Index Object
}

