\docType{class}
\name{RMIndex}
\alias{RMIndex}
\alias{RMIndex-class}
\title{The RMIndex class}
\description{
The \code{RMIndex} object contains Realized Mobility Index
data frame, which is computed for the populations of
Inidividuals.
}
\note{
Object is constructed as a result of rmi.index method call
on the data frame of the Individuals
}
\section{Slots}{
  \describe{ \item{\code{data}:}{Object of class
  \code{"data.frame"}, columns: id - animalId, pop.type -
  population type, ind.home.range - individual home range,
  pop.home.range - population home range, rmi.index - RMI
  value} }
}
\author{
Irina Belyaeva
}

