\name{overlapOverview}
\alias{extractOverlapData}
\alias{overlapOverview}
\title{Provides summary of overlapping time intervals}
\usage{
overlapOverview(dat, start, end, interval, accuracy, minIndiv = 1,
  maxIndiv = NULL, mustIndiv = NULL)

extractOverlapData(dat, possOlaps, rown, completeSetsOnly = TRUE)
}
\arguments{
  \item{dat}{- input data frame}

  \item{dat}{- input data frame}
}
\value{
overlap object

overlap object
}
\description{
Provides summary of overlapping time intervals

Extracts overlaps data from possible overlaps
}
\author{
- Martin Rimmler
}

