\name{summary.MCIndex}
\alias{summary.MCIndex}
\title{Prints Summary for the Movement Coordination Index
This includes statistical tests:
ANOVA
TukeyHSD
Kruskaltest
Kruskalmc}
\usage{
summary.MCIndex(object, ...)
}
\description{
Prints Summary for the Movement Coordination Index This
includes statistical tests: ANOVA TukeyHSD Kruskaltest
Kruskalmc
}

